Update to 2021 release
The new release is not compatible with the image that we previously
used since it requires a more up to date Java version. Instead of
manually updating, we can use the binaries provided which come with
a bundled JRE. The only caveat is that this binary errors out when
run inside Docker, so we use a 3rd-party tool to extract the installer
contents.